requirements.txt
python-dotenv
spacy
openai